// packages/react-native/src/theme/tokens.generated.ts
export const lightTheme: GnomeThemeTokens = {
  accentBgColor: '#3584e4',
  accentColor: '#3584e4',
  accentFgColor: '#fff',
  activeOverlay: 'rgba(0, 0, 0, 0.12)',
  blockquoteBorderWidth: 3,
  blue1: '#99c1f1',
  blue2: '#62a0ea',
  blue3: '#3584e4',
  blue4: '#1c71d8',
  blue5: '#1a5fb4',
  borderSubtle: 'rgba(0, 0, 0, 0.15)',
  brown1: '#cdab8f',
  brown2: '#b5835a',
  brown3: '#986a44',
  brown4: '#865e3c',
  brown5: '#63452c',
  cardBgColor: '#fff',
  cardFgColor: 'rgba(0, 0, 0, 0.8)',
  cardShadeColor: 'rgba(0, 0, 0, 0.12)',
  dark1: '#77767b',
  dark2: '#5e5c64',
  dark3: '#3d3846',
  dark4: '#241f31',
  dark5: '#000',
  destructiveBgColor: '#e01b24',
  destructiveColor: '#e01b24',
  destructiveFgColor: '#fff',
  dialogBackdropBlur: 4,
  dialogBackdropColor: 'rgba(0, 0, 0, 0.5)',
  dialogBgColor: '#fafafa',
  dialogFgColor: 'rgba(0, 0, 0, 0.8)',
  dividerColor: 'rgba(0, 0, 0, 0.07)',
  durationFast: 100,
  durationNormal: 200,
  durationSlow: 400,
  easingDefault: [0.25, 0.46, 0.45, 0.94],
  errorBgColor: '#e01b24',
  errorColor: '#e01b24',
  errorFgColor: '#fff',
  focusRingColor: '#3584e4',
  focusRingOffset: 2,
  focusRingWidth: 3,
  fontFamily: 'Adwaita Sans',
  fontFamilyMono: 'Adwaita Mono',
  fontSizeBody: 16,
  fontSizeCaption: 12,
  fontSizeLargeTitle: 36,
  fontSizeTitle1: 30,
  fontSizeTitle2: 24,
  fontSizeTitle3: 20,
  fontSizeTitle4: 18,
  fontWeightBold: 700,
  fontWeightNormal: 400,
  fontWeightSemibold: 600,
  green1: '#8ff0a4',
  green2: '#57e389',
  green3: '#33d17a',
  green4: '#2ec27e',
  green5: '#26a269',
  headerbarBgColor: '#ebebeb',
  headerbarBorderColor: 'rgba(0, 0, 0, 0.12)',
  headerbarFgColor: 'rgba(0, 0, 0, 0.8)',
  headerbarShadeColor: 'rgba(0, 0, 0, 0.12)',
  hoverOverlay: 'rgba(0, 0, 0, 0.06)',
  layoutBreakpointMedium: 550,
  layoutBreakpointNarrow: 400,
  layoutBreakpointWide: 860,
  layoutContentPadding: 24,
  layoutContentPaddingCompact: 12,
  layoutContentPaddingSpacious: 36,
  layoutSidebarDefaultWidth: 240,
  layoutSidebarMaxWidth: 280,
  layoutSidebarMinWidth: 180,
  layoutSidebarRailWidth: 56,
  light1: '#fff',
  light2: '#f6f5f4',
  light3: '#deddda',
  light4: '#c0bfbc',
  light5: '#9a9996',
  lineHeightBody: 1.5,
  lineHeightHeading: 1.2,
  lineHeightTitle: 1.2,
  newBgColor: '#9141ac',
  newColor: '#9141ac',
  newFgColor: '#fff',
  opacityBorder: 0.15,
  opacityDim: 0.55,
  opacityDisabled: 0.5,
  orange1: '#ffbe6f',
  orange2: '#ffa348',
  orange3: '#ff7800',
  orange4: '#e66100',
  orange5: '#c64600',
  popoverBgColor: '#fff',
  popoverFgColor: 'rgba(0, 0, 0, 0.8)',
  purple1: '#dc8add',
  purple2: '#c061cb',
  purple3: '#9141ac',
  purple4: '#813d9c',
  purple5: '#613583',
  radiusLg: 12,
  radiusMd: 8,
  radiusPill: 9999,
  radiusSm: 4,
  radiusXl: 15,
  red1: '#f66151',
  red2: '#ed333b',
  red3: '#e01b24',
  red4: '#c01c28',
  red5: '#a51d2d',
  sidebarBgColor: '#ebebeb',
  sidebarFgColor: 'rgba(0, 0, 0, 0.8)',
  skeletonBaseColor: 'rgba(0, 0, 0, 0.09)',
  skeletonHighlightColor: 'rgba(255, 255, 255, 0.55)',
  space1: 6,
  space2: 12,
  space3: 18,
  space4: 24,
  space5: 36,
  space6: 48,
  successBgColor: '#2ec27e',
  successColor: '#2ec27e',
  successFgColor: '#fff',
  viewBgColor: '#fff',
  viewFgColor: 'rgba(0, 0, 0, 0.8)',
  warningBgColor: '#f6d32d',
  warningColor: '#e5a50a',
  warningFgColor: 'rgba(0, 0, 0, 0.8)',
  windowBgColor: '#fafafa',
  windowFgColor: 'rgba(0, 0, 0, 0.8)',
  yellow1: '#f9f06b',
  yellow2: '#f8e45c',
  yellow3: '#f6d32d',
  yellow4: '#f5c211',
  yellow5: '#e5a50a',
} as const;

export const darkTheme: GnomeThemeTokens = {
  accentBgColor: '#3584e4',
  accentColor: '#62a0ea',
  accentFgColor: '#fff',
  activeOverlay: 'rgba(255, 255, 255, 0.14)',
  blockquoteBorderWidth: 3,
  blue1: '#99c1f1',
  blue2: '#62a0ea',
  blue3: '#3584e4',
  blue4: '#1c71d8',
  blue5: '#1a5fb4',
  borderSubtle: 'rgba(255, 255, 255, 0.15)',
  brown1: '#cdab8f',
  brown2: '#b5835a',
  brown3: '#986a44',
  brown4: '#865e3c',
  brown5: '#63452c',
  cardBgColor: '#383838',
  cardFgColor: 'rgba(255, 255, 255, 0.87)',
  cardShadeColor: 'rgba(255, 255, 255, 0.12)',
  dark1: '#77767b',
  dark2: '#5e5c64',
  dark3: '#3d3846',
  dark4: '#241f31',
  dark5: '#000',
  destructiveBgColor: '#e01b24',
  destructiveColor: '#e01b24',
  destructiveFgColor: '#fff',
  dialogBackdropBlur: 4,
  dialogBackdropColor: 'rgba(0, 0, 0, 0.65)',
  dialogBgColor: '#242424',
  dialogFgColor: 'rgba(255, 255, 255, 0.87)',
  dividerColor: 'rgba(255, 255, 255, 0.07)',
  durationFast: 100,
  durationNormal: 200,
  durationSlow: 400,
  easingDefault: [0.25, 0.46, 0.45, 0.94],
  errorBgColor: '#e01b24',
  errorColor: '#e01b24',
  errorFgColor: '#fff',
  focusRingColor: '#62a0ea',
  focusRingOffset: 2,
  focusRingWidth: 3,
  fontFamily: 'Adwaita Sans',
  fontFamilyMono: 'Adwaita Mono',
  fontSizeBody: 16,
  fontSizeCaption: 12,
  fontSizeLargeTitle: 36,
  fontSizeTitle1: 30,
  fontSizeTitle2: 24,
  fontSizeTitle3: 20,
  fontSizeTitle4: 18,
  fontWeightBold: 700,
  fontWeightNormal: 400,
  fontWeightSemibold: 600,
  green1: '#8ff0a4',
  green2: '#57e389',
  green3: '#33d17a',
  green4: '#2ec27e',
  green5: '#26a269',
  headerbarBgColor: '#303030',
  headerbarBorderColor: 'rgba(255, 255, 255, 0.1)',
  headerbarFgColor: 'rgba(255, 255, 255, 0.87)',
  headerbarShadeColor: 'rgba(0, 0, 0, 0.36)',
  hoverOverlay: 'rgba(255, 255, 255, 0.08)',
  layoutBreakpointMedium: 550,
  layoutBreakpointNarrow: 400,
  layoutBreakpointWide: 860,
  layoutContentPadding: 24,
  layoutContentPaddingCompact: 12,
  layoutContentPaddingSpacious: 36,
  layoutSidebarDefaultWidth: 240,
  layoutSidebarMaxWidth: 280,
  layoutSidebarMinWidth: 180,
  layoutSidebarRailWidth: 56,
  light1: '#fff',
  light2: '#f6f5f4',
  light3: '#deddda',
  light4: '#c0bfbc',
  light5: '#9a9996',
  lineHeightBody: 1.5,
  lineHeightHeading: 1.2,
  lineHeightTitle: 1.2,
  newBgColor: '#9141ac',
  newColor: '#9141ac',
  newFgColor: '#fff',
  opacityBorder: 0.15,
  opacityDim: 0.55,
  opacityDisabled: 0.5,
  orange1: '#ffbe6f',
  orange2: '#ffa348',
  orange3: '#ff7800',
  orange4: '#e66100',
  orange5: '#c64600',
  popoverBgColor: '#383838',
  popoverFgColor: 'rgba(255, 255, 255, 0.87)',
  purple1: '#dc8add',
  purple2: '#c061cb',
  purple3: '#9141ac',
  purple4: '#813d9c',
  purple5: '#613583',
  radiusLg: 12,
  radiusMd: 8,
  radiusPill: 9999,
  radiusSm: 4,
  radiusXl: 15,
  red1: '#f66151',
  red2: '#ed333b',
  red3: '#e01b24',
  red4: '#c01c28',
  red5: '#a51d2d',
  sidebarBgColor: '#303030',
  sidebarFgColor: 'rgba(255, 255, 255, 0.87)',
  skeletonBaseColor: 'rgba(255, 255, 255, 0.12)',
  skeletonHighlightColor: 'rgba(255, 255, 255, 0.18)',
  space1: 6,
  space2: 12,
  space3: 18,
  space4: 24,
  space5: 36,
  space6: 48,
  successBgColor: '#2ec27e',
  successColor: '#2ec27e',
  successFgColor: '#fff',
  viewBgColor: '#1e1e1e',
  viewFgColor: 'rgba(255, 255, 255, 0.87)',
  warningBgColor: '#f6d32d',
  warningColor: '#e5a50a',
  warningFgColor: 'rgba(0, 0, 0, 0.8)',
  windowBgColor: '#242424',
  windowFgColor: 'rgba(255, 255, 255, 0.87)',
  yellow1: '#f9f06b',
  yellow2: '#f8e45c',
  yellow3: '#f6d32d',
  yellow4: '#f5c211',
  yellow5: '#e5a50a',
} as const;

export const highContrastTheme: GnomeThemeTokens = {
  accentBgColor: '#1a5fb4',
  accentColor: '#1a5fb4',
  accentFgColor: '#fff',
  activeOverlay: 'rgba(0, 0, 0, 0.22)',
  blockquoteBorderWidth: 3,
  blue1: '#99c1f1',
  blue2: '#62a0ea',
  blue3: '#3584e4',
  blue4: '#1c71d8',
  blue5: '#1a5fb4',
  borderSubtle: 'rgba(0, 0, 0, 0.5)',
  brown1: '#cdab8f',
  brown2: '#b5835a',
  brown3: '#986a44',
  brown4: '#865e3c',
  brown5: '#63452c',
  cardBgColor: '#fff',
  cardFgColor: 'rgba(0, 0, 0, 0.8)',
  cardShadeColor: 'rgba(0, 0, 0, 0.12)',
  dark1: '#77767b',
  dark2: '#5e5c64',
  dark3: '#3d3846',
  dark4: '#241f31',
  dark5: '#000',
  destructiveBgColor: '#a51d2d',
  destructiveColor: '#a51d2d',
  destructiveFgColor: '#fff',
  dialogBackdropBlur: 4,
  dialogBackdropColor: 'rgba(0, 0, 0, 0.5)',
  dialogBgColor: '#fafafa',
  dialogFgColor: 'rgba(0, 0, 0, 0.8)',
  dividerColor: 'rgba(0, 0, 0, 0.2)',
  durationFast: 100,
  durationNormal: 200,
  durationSlow: 400,
  easingDefault: [0.25, 0.46, 0.45, 0.94],
  errorBgColor: '#a51d2d',
  errorColor: '#a51d2d',
  errorFgColor: '#fff',
  focusRingColor: '#1a5fb4',
  focusRingOffset: 2,
  focusRingWidth: 3,
  fontFamily: 'Adwaita Sans',
  fontFamilyMono: 'Adwaita Mono',
  fontSizeBody: 16,
  fontSizeCaption: 12,
  fontSizeLargeTitle: 36,
  fontSizeTitle1: 30,
  fontSizeTitle2: 24,
  fontSizeTitle3: 20,
  fontSizeTitle4: 18,
  fontWeightBold: 700,
  fontWeightNormal: 400,
  fontWeightSemibold: 600,
  green1: '#8ff0a4',
  green2: '#57e389',
  green3: '#33d17a',
  green4: '#2ec27e',
  green5: '#26a269',
  headerbarBgColor: '#ebebeb',
  headerbarBorderColor: 'rgba(0, 0, 0, 0.12)',
  headerbarFgColor: 'rgba(0, 0, 0, 0.8)',
  headerbarShadeColor: 'rgba(0, 0, 0, 0.12)',
  hoverOverlay: 'rgba(0, 0, 0, 0.12)',
  layoutBreakpointMedium: 550,
  layoutBreakpointNarrow: 400,
  layoutBreakpointWide: 860,
  layoutContentPadding: 24,
  layoutContentPaddingCompact: 12,
  layoutContentPaddingSpacious: 36,
  layoutSidebarDefaultWidth: 240,
  layoutSidebarMaxWidth: 280,
  layoutSidebarMinWidth: 180,
  layoutSidebarRailWidth: 56,
  light1: '#fff',
  light2: '#f6f5f4',
  light3: '#deddda',
  light4: '#c0bfbc',
  light5: '#9a9996',
  lineHeightBody: 1.5,
  lineHeightHeading: 1.2,
  lineHeightTitle: 1.2,
  newBgColor: '#613583',
  newColor: '#613583',
  newFgColor: '#fff',
  opacityBorder: 0.4,
  opacityDim: 0.7,
  opacityDisabled: 0.6,
  orange1: '#ffbe6f',
  orange2: '#ffa348',
  orange3: '#ff7800',
  orange4: '#e66100',
  orange5: '#c64600',
  popoverBgColor: '#fff',
  popoverFgColor: 'rgba(0, 0, 0, 0.8)',
  purple1: '#dc8add',
  purple2: '#c061cb',
  purple3: '#9141ac',
  purple4: '#813d9c',
  purple5: '#613583',
  radiusLg: 12,
  radiusMd: 8,
  radiusPill: 9999,
  radiusSm: 4,
  radiusXl: 15,
  red1: '#f66151',
  red2: '#ed333b',
  red3: '#e01b24',
  red4: '#c01c28',
  red5: '#a51d2d',
  sidebarBgColor: '#ebebeb',
  sidebarFgColor: 'rgba(0, 0, 0, 0.8)',
  skeletonBaseColor: 'rgba(0, 0, 0, 0.09)',
  skeletonHighlightColor: 'rgba(255, 255, 255, 0.55)',
  space1: 6,
  space2: 12,
  space3: 18,
  space4: 24,
  space5: 36,
  space6: 48,
  successBgColor: '#26a269',
  successColor: '#26a269',
  successFgColor: '#fff',
  viewBgColor: '#fff',
  viewFgColor: 'rgba(0, 0, 0, 0.8)',
  warningBgColor: '#e66100',
  warningColor: '#c64600',
  warningFgColor: 'rgba(0, 0, 0, 0.8)',
  windowBgColor: '#fafafa',
  windowFgColor: 'rgba(0, 0, 0, 0.8)',
  yellow1: '#f9f06b',
  yellow2: '#f8e45c',
  yellow3: '#f6d32d',
  yellow4: '#f5c211',
  yellow5: '#e5a50a',
} as const;